Making proper use of size_t
In a lot of cases, int was being used instead of size_to represent
sizes. That led to some warnings about inconsistencies between int
and size_t usage. While this cl doesn't solve all warnings, it tries
to use size_t and int where it should be appropriate to use them.

Reviewed-on: https://swiftshader-review.googlesource.com/5177
Tested-by: Alexis Hétu <sugoi@google.com>

+++ src/OpenGL/compiler/Intermediate.cpp
@@ -1053,8 +1053,8 @@ bool CompareStruct(const TType& leftNodeType, ConstantUnion* rightUnionArray, Co
     int index = 0;
 
     for (size_t j = 0; j < structSize; j++) {
-        int size = fields[j]->type()->getObjectSize();
-        for (int i = 0; i < size; i++) {
+        size_t size = fields[j]->type()->getObjectSize();
+        for(size_t i = 0; i < size; i++) {
             if (fields[j]->type()->getBasicType() == EbtStruct) {
                 if (!CompareStructure(*(fields[j]->type()), &rightUnionArray[index], &leftUnionArray[index]))
                     return false;
@@ -1078,7 +1078,7 @@ bool CompareStructure(const TType& leftNodeType, ConstantUnion* rightUnionArray,
         int arraySize = leftNodeType.getArraySize();
 
         for (int i = 0; i < arraySize; ++i) {
-            int offset = typeWithoutArrayness.getObjectSize() * i;
+            size_t offset = typeWithoutArrayness.getObjectSize() * i;
             if (!CompareStruct(typeWithoutArrayness, &rightUnionArray[offset], &leftUnionArray[offset]))
                 return false;
         }
@@ -1253,7 +1253,7 @@ ConstantUnion* CreateInverse(TIntermConstantUnion* node, ConstantUnion* unionArr
 TIntermTyped* TIntermConstantUnion::fold(TOperator op, TIntermTyped* constantNode, TInfoSink& infoSink)
 {
     ConstantUnion *unionArray = getUnionArrayPointer();
-    int objectSize = getType().getObjectSize();
+    size_t objectSize = getType().getObjectSize();
 
     if (constantNode) {  // binary operations
         TIntermConstantUnion *node = constantNode->getAsConstantUnion();
@@ -1263,13 +1263,13 @@ TIntermTyped* TIntermConstantUnion::fold(TOperator op, TIntermTyped* constantNod
         // for a case like float f = 1.2 + vec4(2,3,4,5);
         if (constantNode->getType().getObjectSize() == 1 && objectSize > 1) {
             rightUnionArray = new ConstantUnion[objectSize];
-            for (int i = 0; i < objectSize; ++i)
+            for (size_t i = 0; i < objectSize; ++i)
                 rightUnionArray[i] = *node->getUnionArrayPointer();
             returnType = getType();
         } else if (constantNode->getType().getObjectSize() > 1 && objectSize == 1) {
             // for a case like float f = vec4(2,3,4,5) + 1.2;
             unionArray = new ConstantUnion[constantNode->getType().getObjectSize()];
-            for (int i = 0; i < constantNode->getType().getObjectSize(); ++i)
+            for (size_t i = 0; i < constantNode->getType().getObjectSize(); ++i)
                 unionArray[i] = *getUnionArrayPointer();
             returnType = node->getType();
             objectSize = constantNode->getType().getObjectSize();
@@ -1283,14 +1283,14 @@ TIntermTyped* TIntermConstantUnion::fold(TOperator op, TIntermTyped* constantNod
             case EOpAdd:
                 tempConstArray = new ConstantUnion[objectSize];
                 {// support MSVC++6.0
-                    for (int i = 0; i < objectSize; i++)
+                    for (size_t i = 0; i < objectSize; i++)
                         tempConstArray[i] = unionArray[i] + rightUnionArray[i];
                 }
                 break;
             case EOpSub:
                 tempConstArray = new ConstantUnion[objectSize];
                 {// support MSVC++6.0
-                    for (int i = 0; i < objectSize; i++)
+                    for (size_t i = 0; i < objectSize; i++)
                         tempConstArray[i] = unionArray[i] - rightUnionArray[i];
                 }
                 break;
@@ -1300,7 +1300,7 @@ TIntermTyped* TIntermConstantUnion::fold(TOperator op, TIntermTyped* constantNod
             case EOpMatrixTimesScalar:
                 tempConstArray = new ConstantUnion[objectSize];
                 {// support MSVC++6.0
-                    for (int i = 0; i < objectSize; i++)
+                    for (size_t i = 0; i < objectSize; i++)
                         tempConstArray[i] = unionArray[i] * rightUnionArray[i];
                 }
                 break;
@@ -1389,7 +1389,7 @@ TIntermTyped* TIntermConstantUnion::fold(TOperator op, TIntermTyped* constantNod
 					{
 						// Singular matrix, just copy
 						tempConstArray = new ConstantUnion[objectSize];
-						for(int i = 0; i < objectSize; i++)
+						for(size_t i = 0; i < objectSize; i++)
 							tempConstArray[i] = rightUnionArray[i];
 					}
 				}
@@ -1399,7 +1399,7 @@ TIntermTyped* TIntermConstantUnion::fold(TOperator op, TIntermTyped* constantNod
             case EOpIMod:
                 tempConstArray = new ConstantUnion[objectSize];
                 {// support MSVC++6.0
-                    for (int i = 0; i < objectSize; i++) {
+                    for (size_t i = 0; i < objectSize; i++) {
                         switch (getType().getBasicType()) {
                             case EbtFloat:
                                 if (rightUnionArray[i] == 0.0f) {
@@ -1486,7 +1486,7 @@ TIntermTyped* TIntermConstantUnion::fold(TOperator op, TIntermTyped* constantNod
             case EOpLogicalAnd: // this code is written for possible future use, will not get executed currently
                 tempConstArray = new ConstantUnion[objectSize];
                 {// support MSVC++6.0
-                    for (int i = 0; i < objectSize; i++)
+                    for (size_t i = 0; i < objectSize; i++)
                         tempConstArray[i] = unionArray[i] && rightUnionArray[i];
                 }
                 break;
@@ -1494,7 +1494,7 @@ TIntermTyped* TIntermConstantUnion::fold(TOperator op, TIntermTyped* constantNod
             case EOpLogicalOr: // this code is written for possible future use, will not get executed currently
                 tempConstArray = new ConstantUnion[objectSize];
                 {// support MSVC++6.0
-                    for (int i = 0; i < objectSize; i++)
+                    for (size_t i = 0; i < objectSize; i++)
                         tempConstArray[i] = unionArray[i] || rightUnionArray[i];
                 }
                 break;
@@ -1502,7 +1502,7 @@ TIntermTyped* TIntermConstantUnion::fold(TOperator op, TIntermTyped* constantNod
             case EOpLogicalXor:
                 tempConstArray = new ConstantUnion[objectSize];
                 {// support MSVC++6.0
-                    for (int i = 0; i < objectSize; i++)
+                    for (size_t i = 0; i < objectSize; i++)
                         switch (getType().getBasicType()) {
                             case EbtBool: tempConstArray[i].setBConst((unionArray[i] == rightUnionArray[i]) ? false : true); break;
                             default: assert(false && "Default missing");
@@ -1512,51 +1512,51 @@ TIntermTyped* TIntermConstantUnion::fold(TOperator op, TIntermTyped* constantNod
 
             case EOpBitwiseAnd:
                 tempConstArray = new ConstantUnion[objectSize];
-                for(int i = 0; i < objectSize; i++)
+                for(size_t i = 0; i < objectSize; i++)
                     tempConstArray[i] = unionArray[i] & rightUnionArray[i];
                 break;
             case EOpBitwiseXor:
                 tempConstArray = new ConstantUnion[objectSize];
-                for(int i = 0; i < objectSize; i++)
+                for(size_t i = 0; i < objectSize; i++)
                     tempConstArray[i] = unionArray[i] ^ rightUnionArray[i];
                 break;
             case EOpBitwiseOr:
                 tempConstArray = new ConstantUnion[objectSize];
-                for(int i = 0; i < objectSize; i++)
+                for(size_t i = 0; i < objectSize; i++)
                     tempConstArray[i] = unionArray[i] | rightUnionArray[i];
                 break;
             case EOpBitShiftLeft:
                 tempConstArray = new ConstantUnion[objectSize];
-                for(int i = 0; i < objectSize; i++)
+                for(size_t i = 0; i < objectSize; i++)
                     tempConstArray[i] = unionArray[i] << rightUnionArray[i];
                 break;
             case EOpBitShiftRight:
                 tempConstArray = new ConstantUnion[objectSize];
-                for(int i = 0; i < objectSize; i++)
+                for(size_t i = 0; i < objectSize; i++)
                     tempConstArray[i] = unionArray[i] >> rightUnionArray[i];
                 break;
 
             case EOpLessThan:
                 tempConstArray = new ConstantUnion[objectSize];
-                for(int i = 0; i < objectSize; i++)
+                for(size_t i = 0; i < objectSize; i++)
 					tempConstArray[i].setBConst(unionArray[i] < rightUnionArray[i]);
                 returnType = TType(EbtBool, EbpUndefined, EvqConstExpr, objectSize);
                 break;
             case EOpGreaterThan:
 				tempConstArray = new ConstantUnion[objectSize];
-                for(int i = 0; i < objectSize; i++)
+                for(size_t i = 0; i < objectSize; i++)
 					tempConstArray[i].setBConst(unionArray[i] > rightUnionArray[i]);
                 returnType = TType(EbtBool, EbpUndefined, EvqConstExpr, objectSize);
                 break;
             case EOpLessThanEqual:
                 tempConstArray = new ConstantUnion[objectSize];
-                for(int i = 0; i < objectSize; i++)
+                for(size_t i = 0; i < objectSize; i++)
 					tempConstArray[i].setBConst(unionArray[i] <= rightUnionArray[i]);
                 returnType = TType(EbtBool, EbpUndefined, EvqConstExpr, objectSize);
                 break;
             case EOpGreaterThanEqual:
                 tempConstArray = new ConstantUnion[objectSize];
-                for(int i = 0; i < objectSize; i++)
+                for(size_t i = 0; i < objectSize; i++)
 					tempConstArray[i].setBConst(unionArray[i] >= rightUnionArray[i]);
                 returnType = TType(EbtBool, EbpUndefined, EvqConstExpr, objectSize);
                 break;
@@ -1565,7 +1565,7 @@ TIntermTyped* TIntermConstantUnion::fold(TOperator op, TIntermTyped* constantNod
                     if (!CompareStructure(node->getType(), node->getUnionArrayPointer(), unionArray))
                         boolNodeFlag = true;
                 } else {
-                    for (int i = 0; i < objectSize; i++) {
+                    for (size_t i = 0; i < objectSize; i++) {
                         if (unionArray[i] != rightUnionArray[i]) {
                             boolNodeFlag = true;
                             break;  // break out of for loop
@@ -1591,7 +1591,7 @@ TIntermTyped* TIntermConstantUnion::fold(TOperator op, TIntermTyped* constantNod
                     if (CompareStructure(node->getType(), node->getUnionArrayPointer(), unionArray))
                         boolNodeFlag = true;
                 } else {
-                    for (int i = 0; i < objectSize; i++) {
+                    for (size_t i = 0; i < objectSize; i++) {
                         if (unionArray[i] == rightUnionArray[i]) {
                             boolNodeFlag = true;
                             break;  // break out of for loop
@@ -1614,14 +1614,14 @@ TIntermTyped* TIntermConstantUnion::fold(TOperator op, TIntermTyped* constantNod
 			case EOpMax:
                 tempConstArray = new ConstantUnion[objectSize];
                 {// support MSVC++6.0
-                    for (int i = 0; i < objectSize; i++)
+                    for (size_t i = 0; i < objectSize; i++)
                         tempConstArray[i] = unionArray[i] > rightUnionArray[i] ? unionArray[i] : rightUnionArray[i];
                 }
                 break;
             case EOpMin:
                 tempConstArray = new ConstantUnion[objectSize];
                 {// support MSVC++6.0
-                    for (int i = 0; i < objectSize; i++)
+                    for (size_t i = 0; i < objectSize; i++)
                         tempConstArray[i] = unionArray[i] < rightUnionArray[i] ? unionArray[i] : rightUnionArray[i];
                 }
                 break;
@@ -1638,7 +1638,7 @@ TIntermTyped* TIntermConstantUnion::fold(TOperator op, TIntermTyped* constantNod
         //
         TIntermConstantUnion *newNode = 0;
         ConstantUnion* tempConstArray = new ConstantUnion[objectSize];
-        for (int i = 0; i < objectSize; i++) {
+        for (size_t i = 0; i < objectSize; i++) {
             switch(op) {
                 case EOpNegative:
                     switch (getType().getBasicType()) {
@@ -1678,12 +1678,11 @@ TIntermTyped* TIntermConstantUnion::fold(TOperator op, TIntermTyped* constantNod
 
 TIntermTyped* TIntermediate::promoteConstantUnion(TBasicType promoteTo, TIntermConstantUnion* node)
 {
-    int size = node->getType().getObjectSize();
+    size_t size = node->getType().getObjectSize();
 
     ConstantUnion *leftUnionArray = new ConstantUnion[size];
 
-    for (int i=0; i < size; i++) {
-
+    for(size_t i = 0; i < size; i++) {
         switch (promoteTo) {
             case EbtFloat:
                 switch (node->getType().getBasicType()) {
